requirements.txt

Django==1.6.2

Werkzeug==0.9.4

django-extensions==1.3.3
djangorestframework==2.3.12

requests==2.2.1
six==1.5.2
wsgiref==0.1.2

html2text